third_party/WebKit/LayoutTests/fullscreen/api/element-request-fullscreen-and-move-to-iframe-prefixed.html

<!DOCTYPE html>
<title>Element#webkitRequestFullscreen() followed by moving the element into an iframe</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div id="target"></div>
<iframe allowfullscreen></iframe>
<script>
async_test(t => {
  const target = document.getElementById("target");
  const iframeDoc = document.querySelector("iframe").contentDocument;

  iframeDoc.onwebkitfullscreenchange = t.unreached_func("webkitfullscreenchange event in iframe");
  iframeDoc.onwebkitfullscreenerror = t.unreached_func("webkitfullscreenerror event in iframe");
  document.onwebkitfullscreenchange = t.unreached_func("webkitfullscreenchange event");
  document.onwebkitfullscreenerror = t.step_func_done(event => {
    assert_equals(event.target, document);
    assert_equals(document.webkitFullscreenElement, null);
    assert_equals(iframeDoc.webkitFullscreenElement, null);
  });

  trusted_click(t.step_func(() => {
    target.webkitRequestFullscreen();
    iframeDoc.body.appendChild(target);
  }), document.body);
});
</script>
